requirements.txt
annotated-types==0.7.0
anyio==4.11.0
azure-core==1.36.0
azure-storage-blob==12.27.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
fastapi==0.119.1
gunicorn==23.0.0
h11==0.16.0
idna==3.11
isodate==0.7.2
packaging==25.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
